show more text, unjustified, faster
Removed code which formatted and placed one word at a time into one textView at
a time and replaced it with code which concatenates the words into one "synesthesia-
formatted" spannable string, places that in a single large text view then places that in
the relative layout.   This makes it faster and cleaner.